pyproject.toml
[project]
name = "digital-vault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "aiomysql>=0.3.2",
    "fastapi[standard]>=0.135.1",
    "motor>=3.7.1",
    "mypy>=1.19.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "icecream>=2.1.10",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py314"
line-length = 100
include = ["src/**/*.py"]

[tool.ruff.lint]
select = ["B", "B9", "C", "E", "F", "I", "W", "T"]

[tool.mypy]
python_version = "3.14"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
show_error_codes = true
disallow_untyped_defs = false

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]



